extract parseDisplayNameForCorrelation to funnelUtils
